temp_amar_repo/requirements.txt
faiss-cpu>=1.9.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
google-generativeai>=0.3.0
scikit-learn>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
